Include tests for soo DE
